requirements.txt
click==8.1.3
pandas==1.5.2
PyYAML==6.0.1
rich~=12.6.0
scikit_learn==1.1.3
textdistance==4.5.0
tqdm==4.64.1
typer~=0.4.2
pprint